Stripe

+++ zanmi_project/services/queries.py
@@ -1,5 +1,5 @@
 def get_price_in_cents(euro_price: float):
-    return euro_price*100
+    return int(euro_price*100)
